created SetDiggingFeesInput
