+++ src/features/embalm/stepContent/components/DiggingFeesInput.tsx
@@ -1,0 +1,51 @@
+import {
+  Flex,
+  Image,
+  NumberInput,
+  NumberInputField,
+  InputLeftElement,
+  InputGroup,
+  UseNumberInputProps,
+} from '@chakra-ui/react';
+import { removeLeadingZeroes, removeNonIntChars } from 'lib/utils/helpers';
+
+interface DiggingFeesInputProps extends UseNumberInputProps {
+  setDiggingFees: (value: string) => void;
+}
+
+export function DiggingFeesInput({ setDiggingFees, ...rest }: DiggingFeesInputProps) {
+  function handleChangeDiggingFees(valueAsString: string, valueAsNumber: number) {
+    valueAsString = removeNonIntChars(valueAsString);
+    valueAsString = removeLeadingZeroes(valueAsString);
+
+    if (valueAsNumber < 0) {
+      valueAsString = '0';
+      valueAsNumber = 0;
+    }
+
+    if (valueAsString.length > 18) return;
+
+    setDiggingFees(valueAsString);
+  }
+
+  return (
+    <Flex align="center">
+      <InputGroup>
+        <NumberInput
+          w="150px"
+          onChange={handleChangeDiggingFees}
+          {...rest}
+        >
+          <NumberInputField
+            pl={12}
+            pr={1}
+            placeholder="Max"
+          />
+          <InputLeftElement>
+            <Image src="sarco-token-icon.png" />
+          </InputLeftElement>
+        </NumberInput>
+      </InputGroup>
+    </Flex>
+  );
+}
